Very, very dirty implementation of the common api
Just to allow early adopters to play with it.

Obvious problems:

* Non-transactional.
* Loss value types.
* etcd-specific code is in conf/init.lua.
* Common API test cases in etcd test.
* The test uses hardcoded values.
* Only basic test cases are verified.
* No docs.
* Almost no validation.

Aside of this, conf.new() API should be reconsidered. At least it worth
to consider providing ability to use the module itself as a signeton:

 | local conf = require('conf')
 |
 | -- At initialization.
 | conf.cfg({
 |     endpoints = <...>,
 |     <...>
 | })
 |
 | -- When we need a value from the configuration storage.
 | local v = conf.get('foo').data

We should consider ability to reload an application code and maybe even
upgrade the configuration provider itself using some kind of hot reload.

We should make the API versioned from the beginning, otherwise it'll be
hard to enhance it in the future.

Regarding get/set/del: we'll enhance it with support of returning
versions / revisions and to perform preliminary checks about current
version / revision. The first version should at least allow extending
with this kind of arguments and return values.

So, don't take this API as stable in any sense.