Move client code to conf/client/
The main reason of the change is the far plan regarding creating a proxy
module (expose common API via HTTP(S)) and a storage module (store
configuration values in a tarantool replicaset).

The common client is available as `conf.client` and as `conf`, because
(despite future proxy and storage modules) the main part of the module
is the client.

I dropped the driver/ directory, so the etcd client is
`conf.client.etcd` now. The name is descriptive enough, so I don't want
to make it longer than needed (`conf.client.driver.etcd` or so).

I renamed the etcd driver to the etcd client in the API documentation,
because the etcd client is not just code to implement `conf.client`, but
rather useful module per se.