Don't check parent's type on get() and del()
The logic is the following: perform the parent's type check only when
the operation may lead to data inconsistency (if it'll not check the
type). get() is read only, del() just removes a subtree.

In fact, get() should not perform data consistency checks aside of
checks around data it anyway needs. Reductio ad absurdum: imagine that
get() fetch everything and validate each time.